+++ CHANGELOG.md
@@ -3,16 +3,20 @@ CHANGELOG
 
 0.37.0
 ------
+- Added a way to customize the separator of inline info
+  ```sh
+  fzf --info 'inline: ╱ ' --prompt '╱ ' --color prompt:bright-yellow
+  ```
 - New event
     - `focus` - Triggered when the focus changes due to a vertical cursor
       movement or a search result update
       ```sh
-      fzf --bind 'focus:transform-preview-label:echo [ {} ]' --border --preview 'cat {}'
+      fzf --bind 'focus:transform-preview-label:echo [ {} ]' --preview 'cat {}'
 
       # Any action bound to the event runs synchronously and thus can make the interface sluggish
       # e.g. lolcat isn't one of the fastest programs, and every cursor movement in
       #      fzf will be noticeably affected by its execution time
-      fzf --bind 'focus:transform-preview-label:echo [ {} ] | lolcat -f' --border --preview 'cat {}'
+      fzf --bind 'focus:transform-preview-label:echo [ {} ] | lolcat -f' --preview 'cat {}'
 
       # Beware not to introduce an infinite loop
       seq 10 | fzf --bind 'focus:up' --cycle